zeal-rest/src/main/java/com/dzone/albanoj2/zeal/rest/resource/CommentResource.java
package com.dzone.albanoj2.zeal.rest.resource;

/**
 * Resource representing a comment managed by Zeal.
 * 
 * @author Justin Albano <justin.albano.author@gmail.com>
 *
 * @since 1.0.0
 */
public class CommentResource {

	private String id;
	private String articleId;
	private long creationTimestamp;
	private String content;
	
	public String getId() {
		return id;
	}
	
	public void setId(String id) {
		this.id = id;
	}

	public String getArticleId() {
		return articleId;
	}

	public void setArticleId(String articleId) {
		this.articleId = articleId;
	}

	public long getCreationTimestamp() {
		return creationTimestamp;
	}

	public void setCreationTimestamp(long creationTimestamp) {
		this.creationTimestamp = creationTimestamp;
	}

	public String getContent() {
		return content;
	}

	public void setContent(String content) {
		this.content = content;
	}
}
